+++ src/card/model/presentation.rs
@@ -78,6 +78,7 @@ fn object_predicate_implies(predicate: ObjectPredicateDef, expected: ObjectPredi
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::WasDealtDamageThisTurn
+        | ObjectPredicateDef::DealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -129,6 +130,7 @@ fn predicate_color(predicate: ObjectPredicateDef) -> Option<ManaColor> {
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::WasDealtDamageThisTurn
+        | ObjectPredicateDef::DealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -182,6 +184,7 @@ fn predicate_color_count(predicate: ObjectPredicateDef) -> Option<u8> {
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::WasDealtDamageThisTurn
+        | ObjectPredicateDef::DealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -236,6 +239,7 @@ fn predicate_subtype(predicate: ObjectPredicateDef) -> Option<&'static str> {
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::WasDealtDamageThisTurn
+        | ObjectPredicateDef::DealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -292,6 +296,7 @@ fn predicate_negated_subtype(predicate: ObjectPredicateDef) -> Option<&'static s
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::WasDealtDamageThisTurn
+        | ObjectPredicateDef::DealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -354,6 +359,7 @@ fn predicate_power_at_least(predicate: ObjectPredicateDef) -> Option<i16> {
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::WasDealtDamageThisTurn
+        | ObjectPredicateDef::DealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -402,6 +408,7 @@ fn predicate_mana_value_at_most(predicate: ObjectPredicateDef) -> Option<u8> {
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::WasDealtDamageThisTurn
+        | ObjectPredicateDef::DealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -455,6 +462,7 @@ fn predicate_controller(predicate: ObjectPredicateDef) -> Option<PlayerRelation>
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::WasDealtDamageThisTurn
+        | ObjectPredicateDef::DealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -511,6 +519,7 @@ fn predicate_negates(predicate: ObjectPredicateDef, expected: ObjectPredicateDef
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::WasDealtDamageThisTurn
+        | ObjectPredicateDef::DealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource